correct command line parsing for 32-bit platforms

+++ standalone/prep_20220311.f90
@@ -605,7 +605,7 @@ subroutine bounce(varnam,indx,dictionary,ier,mssge)
       integer                                    :: imax
       integer                                    :: i10
 !-----------------------------------------------------------------------------------------------------------------------------------
-      maxtry=int(log(float(IPic))/log(2.0)+1.0)                 ! calculate max number of tries required to find a conforming name
+      maxtry=nint(log(float(IPic))/log(2.0)+1.0)                 ! calculate max number of tries required to find a conforming name
       indx=(IPic+1)/2
       imin=1
       imax=IPic
@@ -15978,7 +15978,7 @@ integer                                 :: error
       maxtry=0
       place=-1
    else
-      maxtry=int(log(float(arraysize))/log(2.0)+1.0)
+      maxtry=nint(log(float(arraysize))/log(2.0)+1.0)
       place=(arraysize+1)/2
    endif
    imin=1
@@ -16058,7 +16058,7 @@ integer                                :: error
       maxtry=0
       place=-1
    else
-      maxtry=int(log(float(arraysize))/log(2.0)+1.0)
+      maxtry=nint(log(float(arraysize))/log(2.0)+1.0)
       place=(arraysize+1)/2
    endif
    imin=1
@@ -16138,7 +16138,7 @@ integer                                :: error
       maxtry=0
       place=-1
    else
-      maxtry=int(log(float(arraysize))/log(2.0)+1.0)
+      maxtry=nint(log(float(arraysize))/log(2.0)+1.0)
       place=(arraysize+1)/2
    endif
    imin=1
@@ -16218,7 +16218,7 @@ integer                                :: error
       maxtry=0
       place=-1
    else
-      maxtry=int(log(float(arraysize))/log(2.0)+1.0)
+      maxtry=nint(log(float(arraysize))/log(2.0)+1.0)
       place=(arraysize+1)/2
    endif
    imin=1
